requirements.txt
pyyaml
dvc~=3.8
kaggle~=1.5.13
pandas[performance]~=2.0.1
more_itertools
numpy~=1.20
ipykernel
hvplot~=0.8.3
matplotlib~=3.7
scipy~=1.10.0
plotly~=5.15.0
implicit~=0.6.1
ipywidgets~=8.0
flatten-dict~=0.4.2
seaborn~=0.12.2
lightfm~=1.17.0
catboost~=1.2.0
numba~=0.57.0
python-dotenv~=1.0.0
hydra-core~=1.3.2
scikit-learn~=1.3.0
hydra-optuna-sweeper~=1.2.0
nbformat>=4.2.0
polars[pandas,numpy,timezone,pyarrow]~=0.17.14
shap~=0.42.1
fasttext~=0.9.2
huggingface-hub~=0.16.4
ydata-profiling~=4.5.1
wandb~=0.15.8


